backend/requirements.txt
Flask==3.0.3
Flask_Cors==4.0.1
Jinja2==3.1.2
openai==1.55.3
PyYAML==6.0.1
Requests==2.32.3
gitpython==3.1.43
watchtower==3.2.0
marshmallow==3.21.3
python-dotenv==1.0.1
boto3==1.34.144
sentry-sdk==2.13.0
flask-executor==1.0.0
langchain==0.2.16
langchain_aws==0.1.18
